Cleanup flow in DeliverTx and CheckTx
